refactor base resource
