+++ app/controllers/avo/attachments_controller.rb
@@ -8,7 +8,7 @@ module Avo
 
     def create
       blob = ActiveStorage::Blob.create_and_upload! io: params[:file].to_io, filename: params[:filename]
-      association_name = BaseResource.valid_attachment_name(@record, params[:attachment_key])
+      association_name = Resources::Base.valid_attachment_name(@record, params[:attachment_key])
 
       if association_name.blank?
         raise ActionController::BadRequest.new("Could not find the attachment association for #{params[:attachment_key]} (check the `attachment_key` for this Trix field)")
