+++ app/controllers/avo/search_controller.rb
@@ -88,9 +88,9 @@ module Avo
       # to scope the query.
       # Example usage: Got to a project, create a new review, and search for a user.
       if parent.blank? && params[:via_parent_resource_id].present? && params[:via_parent_resource_class].present? && params[:via_relation].present?
-        parent_resource_class = BaseResource.get_model_by_name params[:via_parent_resource_class]
+        parent_resource_class = Resources::Base.get_model_by_name params[:via_parent_resource_class]
 
-        reflection_class = BaseResource.get_model_by_name params[:via_reflection_class]
+        reflection_class = Resources::Base.get_model_by_name params[:via_reflection_class]
 
         grandparent = parent_resource_class.find params[:via_parent_resource_id]
         parent = reflection_class.new(
@@ -103,7 +103,7 @@ module Avo
 
     # This scope is applied if the search is being performed on a has_many association
     def apply_has_many_scope
-      association_name = BaseResource.valid_association_name(parent, params[:via_association_id])
+      association_name = Resources::Base.valid_association_name(parent, params[:via_association_id])
 
       # Get association records
       query = parent.send(association_name)
