+++ lib/generators/avo/templates/resource/resource.tt
@@ -1,4 +1,4 @@
-class Avo::Resources::<%= resource_class %> < Avo::BaseResource
+class Avo::Resources::<%= resource_class %> < Avo::Resources::Base
   self.includes = []<%= model_class_from_args %>
   # self.search = {
   #   query: -> { query.ransack(id_eq: params[:q], m: "or").result(distinct: false) }
