+++ spec/dummy/app/avo/resources/city.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class Avo::Resources::City < Avo::BaseResource
+class Avo::Resources::City < Avo::Resources::Base
   self.search = {
     query: -> { query.ransack(id_eq: params[:q], m: "or").result(distinct: false) }
   }
