+++ spec/dummy/app/avo/resources/event.rb
@@ -1,4 +1,4 @@
-class Avo::Resources::Event < Avo::BaseResource
+class Avo::Resources::Event < Avo::Resources::Base
   self.title = :name
   self.description = "An event that happened at a certain time."
   self.includes = [:location]
