+++ spec/dummy/app/avo/resources/sibling.rb
@@ -1,4 +1,4 @@
-class Avo::Resources::Sibling < Avo::BaseResource
+class Avo::Resources::Sibling < Avo::Resources::Base
   self.description = "Demo resource to illustrate Avo\'s Single Table Inheritance support (Sibling < Person)"
   self.model_class = ::Sibling
   self.search = {
